refactor nconf layout
